Refactor v2 validator dependencies
